Add a main() function to be called from katsdpcontim
To allow the continuum pipeline to run the QA validation code without calling a separate process, the code is reorganised to be called from a main function.

Additionally Aegean and Bane are now run using the `subprocess` module which allows the validation code to connect to their `stdout` and `stderr` pipes.