Limit the source count QA metric to > 500uJy
A QA metric is generated using the source counts of the extracted
sources. This is the reduced chi squared value of the Norris et al.,
2011 polynomial fit to the extracted source counts.

As the imager cleans to a threshold of 100uJy/beam (in apparent sky) it
is unlikely that the fluxes of fainter sources are reliable. Therefore
it is prudent to restrict this metric to the brighter sources (> 500uJy)
in the field only.